requirements.txt
botocore==1.7.47
certifi==2017.11.5
chardet==3.0.4
click==6.7
docutils==0.14
idna==2.6
jmespath==0.9.3
python-dateutil==2.6.1
python-dotenv==0.7.1
requests==2.18.4
s3transfer==0.1.11
selenium==3.7.0
six==1.11.0
urllib3==1.22
